+++ api/app/Http/Controllers/API/CalentamientoController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\CalentamientoResource;
 use App\Models\Calentamiento;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CalentamientoController extends Controller
 {
@@ -16,6 +17,7 @@ class CalentamientoController extends Controller
     {
         try {
 
+            // devuelve el recurso, ordenado por id
             $calentamientos = CalentamientoResource::collection(
                 Calentamiento::orderBy('nombre')->get(),
             );
@@ -38,10 +40,16 @@ class CalentamientoController extends Controller
     {
         try {
 
+            Gate::authorize('create', Calentamiento::class);
+
+            // obtiene la información de $request y la convierte a un array asociativo
             $calentamiento = json_decode($request->getContent(), true);
+
+            // crea el calentamiento
             $calentamiento = Calentamiento::create($calentamiento);
 
             return response()->json([
+                'message' => 'Calentamiento creado.',
                 'data' => new CalentamientoResource($calentamiento),
             ], 201);
 
@@ -55,6 +63,7 @@ class CalentamientoController extends Controller
      */
     public function show($id)
     {
+        // encuentra el calentamiento
         $calentamiento = Calentamiento::findOrFail($id);
 
         return response()->json([
@@ -67,21 +76,26 @@ class CalentamientoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('update', Calentamiento::class);
+
+        // encuentra el calentamiento
         $calentamiento = Calentamiento::find($id);
 
         if ($calentamiento) {
 
+            // lo valida
             $request->validate([
                 'nombre' => 'required',
                 'descripcion' => 'required',
             ]);
 
+            // actualiza valores
             $calentamiento->nombre = $request->input('nombre');
             $calentamiento->descripcion = $request->input('descripcion');
             $calentamiento->imagen = $request->input('imagen');
             $calentamiento->save();
 
-            return response('', 204);
+            return response()->json(['message' => 'Calentamiento actualizado.'], 200);
 
         } else {
             return response()->json([
@@ -95,12 +109,17 @@ class CalentamientoController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete', Calentamiento::class);
+
+        // encuentra el calentamiento
         $calentamiento = Calentamiento::find($id);
 
         if ($calentamiento) {
 
+            // lo elimina
             $calentamiento->delete();
-            return response('', 204);
+
+            return response()->json(['message' => 'Calentamiento eliminado.'], 200);
 
         } else {
             return response()->json([
