+++ api/app/Http/Controllers/API/EjercicioController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\EjercicioResource;
 use App\Models\Ejercicio;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EjercicioController extends Controller
 {
@@ -16,6 +17,9 @@ class EjercicioController extends Controller
     {
         try {
 
+            Gate::authorize('viewAny', Ejercicio::class);
+
+            // devuelve el recurso ordenado por id
             $ejercicios = EjercicioResource::collection(
                 Ejercicio::orderBy('id')->get(),
             );
@@ -36,11 +40,16 @@ class EjercicioController extends Controller
     {
         try {
 
+            Gate::authorize('create', Ejercicio::class);
+
             // obtiene la información de $request y la convierte a un array asociativo
             $ejercicio = json_decode($request->getContent(), true);
+
+            // crea el ejercicio
             $ejercicio = Ejercicio::create($ejercicio);
 
             return response()->json([
+                'message' => 'Ejercicio creado.',
                 'data' => new EjercicioResource($ejercicio),
             ], 201);
 
@@ -54,6 +63,9 @@ class EjercicioController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('view', Ejercicio::class);
+
+        // encuentra el ejercicio
         $ejercicio = Ejercicio::findOrFail($id);
 
         return response()->json([
@@ -66,21 +78,26 @@ class EjercicioController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('update', Ejercicio::class);
+
+        // encuentra el ejercicio
         $ejercicio = Ejercicio::findOrFail($id);
 
+        // valida campos
         $request->validate([
             'nombre' => 'required',
             'descripcion' => 'required',
             'grupo_id' => 'required | exists:grupos_musculares,id',
         ]);
 
+        // los actualiza
         $ejercicio->nombre = $request->input('nombre');
         $ejercicio->descripcion = $request->input('descripcion');
         $ejercicio->imagen = $request->input('imagen');
         $ejercicio->grupo_id = $request->input('grupo_id');
         $ejercicio->save();
 
-        return response('', 204);
+        return response()->json(['message' => 'Ejercicio actualizado.'], 200);
     }
 
     /**
@@ -88,9 +105,14 @@ class EjercicioController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete', Ejercicio::class);
+
+        // encuentra el ejercicio
         $ejercicio = Ejercicio::findOrFail($id);
 
+        // lo elimina
         $ejercicio->delete();
-        return response('', 204);
+
+        return response()->json(['message' => 'Ejercicio eliminado.'], 200);
     }
 }
