+++ api/app/Http/Controllers/API/EntrenadorController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\EntrenadorResource;
 use App\Models\Entrenador;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EntrenadorController extends Controller
 {
@@ -16,6 +17,9 @@ class EntrenadorController extends Controller
     {
         try {
 
+            Gate::authorize('viewAny', Entrenador::class);
+
+            // devaulve el recurso ordenado por id
             $entrenadores = EntrenadorResource::collection(
                 Entrenador::orderBy('id')->get(),
             );
@@ -36,8 +40,12 @@ class EntrenadorController extends Controller
     {
         try {
 
+            Gate::authorize('create', Entrenador::class);
+
             // obtiene la información de $request y la convierte a un array asociativo
             $entrenador = json_decode($request->getContent(), true);
+
+            // crea el entrenador
             $entrenador = Entrenador::create($entrenador);
 
             return response()->json([
@@ -54,9 +62,13 @@ class EntrenadorController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('view', Entrenador::class);
+
+        // encuentra el entrenador
         $entrenador = Entrenador::findOrFail($id);
 
         return response()->json([
+            'message' => 'Entrenador creado.',
             'data' => new EntrenadorResource($entrenador),
         ], 200);
     }
@@ -66,8 +78,12 @@ class EntrenadorController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // encuentra el entrenador
         $entrenador = Entrenador::findOrFail($id);
 
+        Gate::authorize('update', [Entrenador::class, $entrenador]);
+
+        // valida campos
         $request->validate([
             'nombre' => 'required',
             'apellidos' => 'required',
@@ -76,6 +92,7 @@ class EntrenadorController extends Controller
             'user_id' => 'required | exists:users,id'
         ]);
 
+        // los actualiza
         $entrenador->nombre = $request->input('nombre');
         $entrenador->apellidos = $request->input('apellidos');
         $entrenador->email = $request->input('email');
@@ -83,7 +100,7 @@ class EntrenadorController extends Controller
         $entrenador->user_id = $request->input('user_id');
         $entrenador->save();
 
-        return response('', 204);
+        return response()->json(['message' => 'Entrenador actualizado.'], 200);
     }
 
     /**
@@ -91,9 +108,14 @@ class EntrenadorController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete', Entrenador::class);
+
+        // encuentra el entrenador
         $entrenador = Entrenador::findOrFail($id);
 
+        // lo elimina
         $entrenador->delete();
-        return response('', 204);
+
+        return response()->json(['message' => 'Entrenador eliminado.'], 200);
     }
 }
