+++ api/app/Http/Controllers/API/ProductoController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ProductoResource;
 use App\Models\Producto;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductoController extends Controller
 {
@@ -16,6 +17,7 @@ class ProductoController extends Controller
     {
         try {
 
+            // devuelve recurso ordenado por id
             $productos = ProductoResource::collection(
                 Producto::orderBy('id')->get(),
             );
@@ -36,8 +38,12 @@ class ProductoController extends Controller
     {
         try {
 
+            Gate::authorize('create', Producto::class);
+
             // obtiene la información de $request y la convierte a un array asociativo
             $producto = json_decode($request->getContent(), true);
+
+            // crea el producto
             $producto = Producto::create($producto);
 
             return response()->json([
@@ -54,9 +60,11 @@ class ProductoController extends Controller
      */
     public function show($id)
     {
+        // encuentra el producto
         $producto = Producto::findOrFail($id);
 
         return response()->json([
+            'message' => 'Producto creado.',
             'data' => new ProductoResource($producto),
         ], 200);
     }
@@ -66,20 +74,25 @@ class ProductoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('update', Producto::class);
+
+        // encuentra el producto
         $producto = Producto::findOrFail($id);
 
+        // valida campos
         $request->validate([
             'nombre' => 'required',
             'descripcion' => 'required',
             'precio' => 'required',
         ]);
 
+        // los actualiza
         $producto->nombre = $request->input('nombre');
         $producto->descripcion = $request->input('descripcion');
         $producto->precio = $request->input('precio');
         $producto->save();
 
-        return response('', 204);
+        return response()->json(['message' => 'Producto actualizado.'], 200);
     }
 
     /**
@@ -87,9 +100,14 @@ class ProductoController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete', Producto::class);
+
+        // encuentra el producto
         $producto = Producto::findOrFail($id);
 
+        // lo elimina
         $producto->delete();
-        return response('', 204);
+
+        return response()->json(['message' => 'Producto eliminado.'], 200);
     }
 }
