+++ api/app/Http/Controllers/API/TaquillaController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\TaquillaResource;
 use App\Models\Taquilla;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TaquillaController extends Controller
 {
@@ -16,6 +17,9 @@ class TaquillaController extends Controller
     {
         try {
 
+            Gate::authorize('viewAny', Taquilla::class);
+
+            // devuelve el recurso ordenado por id
             $taquillas = TaquillaResource::collection(
                 Taquilla::orderBy('id')->get(),
             );
@@ -36,8 +40,12 @@ class TaquillaController extends Controller
     {
         try {
 
+            Gate::authorize('create', Taquilla::class);
+
             // obtiene la información de $request y la convierte a un array asociativo
             $taquilla = json_decode($request->getContent(), true);
+
+            // crea la taquilla
             $taquilla = Taquilla::create($taquilla);
 
             return response()->json([
@@ -54,9 +62,13 @@ class TaquillaController extends Controller
      */
     public function show($id)
     {
+        // encuentra la taquilla
         $taquilla = Taquilla::findOrFail($id);
 
+        Gate::authorize('view', [Taquilla::class, $taquilla]);
+
         return response()->json([
+            'message' => 'Taquilla creada.',
             'data' => new TaquillaResource($taquilla),
         ], 200);
     }
@@ -66,18 +78,23 @@ class TaquillaController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('update', Taquilla::class);
+
+        // encuentra la taquilla
         $taquilla = Taquilla::findOrFail($id);
 
+        // valida los campos
         $request->validate([
             //'socio_id' => 'required',
             //'fecha_fianza' => 'required',
         ]);
 
+        // los actualiza
         $taquilla->socio_id = $request->input('socio_id');
         $taquilla->fecha_fianza = $request->input('fecha_fianza');
         $taquilla->save();
 
-        return response('', 204);
+        return response()->json(['message' => 'Taquilla actualizada.'], 200);
     }
 
     /**
@@ -85,9 +102,14 @@ class TaquillaController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete', Taquilla::class);
+
+        // encuentra la taquilla
         $taquilla = Taquilla::findOrFail($id);
 
+        // la elimina
         $taquilla->delete();
-        return response('', 204);
+
+        return response()->json(['message' => 'Taquilla eliminada.'], 200);
     }
 }
